fixed build reorganized file locations
