+++ src/main/java/sysc4806/pm4y/controllers/CoordinatorController.java
@@ -1,12 +1,13 @@
 package sysc4806.pm4y.controllers;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
-import sysc4806.pm4y.models.*;
-import java.util.*;
-import org.springframework.*;
-import org.springframework.stereotype.Controller;
+import sysc4806.pm4y.models.ProjectCoordinator;
+import sysc4806.pm4y.repositories.ProjectRepo;
+import sysc4806.pm4y.repositories.UserRepo;
 
 @Controller
 public class CoordinatorController
