+++ src/main/java/sysc4806/pm4y/controllers/IndexController.java
@@ -4,8 +4,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import javax.xml.ws.RequestWrapper;
-
 @Controller
 public class IndexController {
