+++ src/main/java/sysc4806/pm4y/controllers/RequestController.java
@@ -3,7 +3,6 @@ package sysc4806.pm4y.controllers;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
 public class RequestController {
