+++ src/main/java/sysc4806/pm4y/models/Student.java
@@ -1,13 +1,17 @@
 package sysc4806.pm4y.models;
 
+import javax.persistence.CascadeType;
 import javax.persistence.Entity;
-import java.util.*;
+import javax.persistence.OneToOne;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 public class Student extends User
 {
+    @OneToOne(targetEntity = Student.class, cascade = CascadeType.ALL)
+    private List<Student> group = new ArrayList<>();
     private boolean hasgroup;
-    private List<Student> group = new ArrayList<Student>();
     public Student(final String first,final String last)
     {
         super(first, last);
